pyproject.toml
[tool.poetry]
name = "discord-bot-3"
version = "0.1.0"
description = ""
authors = ["Matt Wood <matt@cratebind.com>"]

[tool.poetry.dependencies]
python = "3.9.0"
"discord.py" = "^1.5.1"
requests = "^2.25.0"
python-dotenv = "^0.15.0"
tweepy = "^3.9.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
